signlink: Move signlink out to a shared package
